Correctly build Icu when cross-compiling.
When cross-compiling, icu need to be build two times :
- Once in 'native' mode.
- Once cross-compiling.

This is needed as the cross-compiling need to launch executable that are
build in the first step.

Icu dependency become a bit more complex as the second compilation needs to
reference the first compilation and that configure options and buildEnv
change.