Do not depends on custom fedora tools for cross-compile.
Fedora provides a set of (powerfull) tools to correctly pass the right
options to compilation tools (configure, cmake, ...)

This is great but we must not use them as we want to cross-compile from
other distribution (ubuntu).

Now, we do not use those tools and we can easily port cross-compilation to
ubuntu.