Fix installation of meson and pillow (pip3) in travis.
I don't know why, but the travis build doesn't work since few days.
It seems that pip3 cannot install in `/usr/local` anymore.
Probably related to last travis change* as a lot of people seems to have
problem with this update and python (even if this is not the same problems)

Anyway, we should not install a dependencies in global environment without
a sudo. Let's install the pip3 dependencies for the user only.

[*] https://docs.travis-ci.com/user/build-environment-updates/2017-09-06/