Force the source_dir of ctpp2 source.
The `source_dir` is automatically infer from the name of the dependency.
As the ctpp2 source is use in two dependencies (ctpp2 and ctpp2c), if
we use the default source_dir, kiwix-build will use one name and the
other dependency will not find the source directory.

By forcing the `source_dir`, both dependencies will use the same directory.