We must run the deploy command even on tag push.
In case of tag push, travis will trigger a build but will set
TRAVIS_BRANCH to the tag name, not "master".
So by default, the deploy script will be skipped in case of tag push.

There is a option "tags: true" to deploy on tag push. But it will run
deploy script ONLY on tag push, not on cron build.

So we need two "identical" deploy script, one for tags only and
one for cron build.

As we also need to activate travis build on push (because we push tags),
travis will end to build twice a PR (one for the PR and one for the pushed
branch). So we need to allow build only for `master` branches and tags.