Try to decrypt the secret key only if we will do a deployement.
Is a contributor do a PR from its own repository, we will try do decrypt
the key but fail because secret env var are not available.

We need the secret key only if we do a deployement and the deployement is
done only for a tag or a cron job.