Do not try to decrypt the ssh_key if we cannot.
If we are building a PR comming from a fork, the secret key will not
be available and we cannot decrypt the ssh key.

And if the ssh key cannot be decrypted, we should not try to use it to
upload base dependencies.

`cron` and deployement are made only on "secured environment" so we don't
need to test for existance of the key because we are sure that we have
been able to decrypt it.