[Travis] Allow the desktop appimage build to fail.
`linuxqtdeploy` recently add a test on the running system (libc version)
and refuse to run if the libc is too recent.
So we would need to run `linuxqtdeploy` only on ubuntu 14.04.
However, we also need `qtwebengine5-dev` package whom is available only
since ubuntu 18.04.

While we find a solution (see comments in the linuxdeployqt issue), we
have to allow the appimage build to fail.

See probonopd/linuxdeployqt#340