Do not strip libraries in release_mode.
Meson use gcc's `-s` option to create the archive.
This option should make gcc meemic `ar` tool but in some version of gcc,
it seems there is a bug and the `-s` option is not taken into account.

See https://github.com/mesonbuild/meson/issues/4138

So, do not try to strip libraries. We will strip the final binaries anyway.