Do not include low level std lib in appimage package
We have added those libraries to fix kiwix/kiwix-destop#14,
ie allow appimage build on recent system to run on older ones by packaging
the libc and libstdc++ library in the appimage.

But by packaging libstdc++, we force kiwix-desktop to use the packaged one.
On some recent system, the mesa drivers need a recent version of libstdc++ and,
as we are force to use the packaged one, we may use a "old" one.

The workaround forcing a driver was bypassing the problem by forcing a driver
not using a recent libstdc++. But with time, it is more and more difficult
to find one.

This commit means that new appimage generated will works only on equal or
more recent system that the one on which it has been generated.
For now we generate the appimage on Ubuntu bionic (18.04 LTS).

This reverts commit 05c0a691ec407a6ae53d9b8f405bce90e119ae3e.