Force the rebuild of kiwix-desktop on windows with correct dependencies.
On appveyor we always use the last version of the master.
This is ok when we build a release just after creating it but here
we need to build kiwix-desktop with the last release of libkiwix, master
is buggy.

Do not merge with master as it will freeze the libkiwix and libzim version.