Better definition of what to build in the CI.
The current way to define what to build can be error prone as it is a
succession of `if`.

This commit tries to fix that by introducing a table of what to build.

The definition algorithm is based on the *visual* representation of the
table. It works this way:
1. Reduce the column range (Range) by looping on the header lines.
2. Once the final range is found, loop on the project lines and resolve
   what to build.

1.
Each line before the `======` is a header line.
Line is divided in cells (separated by `|`) .
The first cell is a key to lookup in a context.
Other cells are "value" to match with the context.
A cell's value match, we reduce the column range and go the next line.
The final range is the "context range".

"values" are :
- A empty cell: We always match (equivalent to `re'.*`)
- A string starting by `re'`: What follow is a regex to make match.
- A string starting by `eval'`: What follow is a string to eval before...
- A string (or a value if `eval') which must be equal to the context.

2.
Each line after the `======` is a header line.
Line is divided in cells (separated by `|`) .
The first cell is the name of the project to build (or not).
Other cells are "build definition" for the current project.
We match the cells' range with the "context range".
For each cell matching the "context range" we apply the "build definition".
For now, "build definition" is `x` (build) or nothing.

Please not that cells can span across "previous lines" as the main
information is the column range and cells are just use to reduce it.