Fix cross-compilation to raspberry-pi.
Use a toolschain made for debian Bullseye (or equivalent).
Using the Stretch toolschain, kiwix-serve launch but crash with a SIGSEGV
at first request.
(I suspect a "conflict" between thread and exception handling as
a test program with exception only works).

Moving the a toolschain targeting Bullseye fixes the issues
(both on qemu and real raspberry-pi v3 B).

Note that the toolschain is given to compile for Pi 2 and 3 only.
I've also tested with toolschain for Pi 3A+, 3B+, 4 and generated binary
also works.
Not tested of Pi Zero.


Fix #613