Upgraded Android NDK to r23c
Android NDK r23c doesn't work out of the box because of some bugs in the
prebuilt toolchain - certain entries in the bin directory, instead of
being symlinks, are regular text files containing the name of the
sibling file that they should be referring to. Hence the need to patch
the make_standalone_toolchain.py script.