Fixed deprecation warnings by meson
Fixed the following deprecation warnings issued by meson 1.6.1

DEPRECATION: "pkgconfig" entry is deprecated and should be replaced by "pkg-config"
DEPRECATION: c_link_args in the [properties] section of the machine file is deprecated, use the [built-in options] section.
DEPRECATION: cpp_link_args in the [properties] section of the machine file is deprecated, use the [built-in options] section.
DEPRECATION: c_args in the [properties] section of the machine file is deprecated, use the [built-in options] section.
DEPRECATION: cpp_args in the [properties] section of the machine file is deprecated, use the [built-in options] section.